Move SPDX license check to `spdx_common`
Previously `is_spdx_license_expression` method was duplicated in
both tag/value and json package_helpers. This change moves it to
`spdx_common` for general use
