+++ tern/formats/spdx/spdxjson/package_helpers.py
@@ -7,19 +7,9 @@
 Helper functions for packages in SPDX JSON document creation
 """
 
-from license_expression import get_spdx_licensing
 from tern.report import content
 from tern.formats.spdx import spdx_common
 from tern.formats.spdx.spdxjson import formats as json_formats
-
-
-def is_spdx_license_expression(license_data):
-    '''Return True if the license is a valid SPDX license expression, else
-    return False'''
-    licensing = get_spdx_licensing()
-    if ',' in license_data:
-        license_data = license_data.replace(',', ' ')
-    return licensing.validate(license_data).errors == []
 
 
 def get_package_comment(package):
@@ -35,7 +25,7 @@ def get_package_comment(package):
 
 def get_package_license_declared(package_license_declared):
     if package_license_declared:
-        if is_spdx_license_expression(package_license_declared):
+        if spdx_common.is_spdx_license_expression(package_license_declared):
             return package_license_declared
         return spdx_common.get_license_ref(package_license_declared)
     return 'NONE'
